Make TsWriter compile. Will do my rework in here later. Just need something we can test/release!

+++ DirectShowFilters/TsWriter/source/BatParser.h
@@ -30,7 +30,7 @@ class CBatParser : public CNitParser
   public:
     CBatParser(void);
     virtual ~CBatParser(void);
-    void GetBouquetIds(int originalNetworkId, int transportStreamId, int serviceId, vector<int>* bouquetIds);
+    void GetBouquetIds(int originalNetworkId, int transportStreamId, int serviceId, vector<unsigned short>* bouquetIds);
     int GetBouquetNameCount(int bouquetId);
     void GetBouquetName(int bouquetId, int index, unsigned int* language, char** name);
 };
